Truncate to subtexel for linear. Handle pixel center mapping to texel center for linear.

+++ pcsx2/GS/Renderers/DX12/GSDevice12.cpp
@@ -3217,7 +3217,7 @@ const ID3DBlob* GSDevice12::GetTFXPixelShader(const GSHWDrawConfig::PSSelector& 
 	sm.AddMacro("PS_AA1", static_cast<u32>(sel.aa1));
 	sm.AddMacro("PS_ABE", sel.abe);
 	sm.AddMacro("PS_ANISOTROPIC_FILTERING", sel.sw_aniso);
-	sm.AddMacro("PS_ROUND_UV", sel.round_uv);
+	sm.AddMacro("PS_ROUND_UV", static_cast<u32>(sel.round_uv));
 
 	ComPtr<ID3DBlob> ps(m_shader_cache.GetPixelShader(m_tfx_source, sm.GetPtr(), "ps_main"));
 	it = m_tfx_pixel_shaders.emplace(sel, std::move(ps)).first;
